+++ index.ts
@@ -4,3 +4,4 @@ export { gcd } from "./src/gcd.js";
 export { combination } from "./src/combination.js";
 export { lcm } from "./src/lcm.js";
 export { permutation } from "./src/permutation.js";
+export { modPow } from "./src/mod_pow.js";